// pslpwcl/src/app/modules/pslpapi/model/conoscenzaInformatica.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Decodifica } from './decodifica';
import { LavInformatica } from './lavInformatica';

export interface ConoscenzaInformatica { 
    certificato?: boolean;
    conoscenzaSpecifica?: Decodifica;
    gradoConoscenza?: Decodifica;
    lavInformatica?: LavInformatica;
    modalitaApprendimento?: Decodifica;
    note?: string;
    tipoConoscenza?: Decodifica;
}