// pslpwcl/src/app/modules/pslpapi/model/lavInformatica.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Fontedato } from './fontedato';
import { GradoConInf } from './gradoConInf';
import { InformaticaDett } from './informaticaDett';
import { LavAnagrafica } from './lavAnagrafica';
import { ModApprInf } from './modApprInf';

export interface LavInformatica { 
    codUserAggiorn?: string;
    codUserInserim?: string;
    dataAggiorn?: Date;
    dataInserim?: Date;
    dsNote?: string;
    flgCertificato?: string;
    idSilLavInformatica?: number;
    silLavAnagrafica?: LavAnagrafica;
    silTFontedato?: Fontedato;
    silTGradoConInf?: GradoConInf;
    silTInformaticaDett?: InformaticaDett;
    silTModApprInf?: ModApprInf;
    timestamp?: number;
}